chapter-7/07_04/azure-function/requirements.txt
azure-functions
azure-identity
azure-storage-file-share
azure-storage-file
azure-storage-blob
requests